Implement CsmpAgentLib_sample_util header and source files
