+++ sample/linux.target
@@ -10,7 +10,7 @@ LIBS += -lpthread
 #LIBS += -lcrypto
 #LIBS += -ldl
 
-OBJ = CsmpAgentLib_sample.o signature_verify.o
+OBJ = CsmpAgentLib_sample.o CsmpAgentLib_sample_util.o signature_verify.o
 
 LIB_OBJECT = csmp_agent_lib.a
 OBJECT = CsmpAgentLib_sample
